requirements.txt
redis==5.0.3
psycopg==3.1.18
psycopg-binary==3.1.18
Django==5.0.3
celery==5.3.6
django-celery-beat==2.6.0
django-guardian==2.4.0
django-redis==5.4.0
django-timezone-field==6.1.0
django-unfold==0.21.1
django-celery-results==2.5.1
websockets==12.0
requests==2.31.0
django-cors-headers==4.3.1
djangorestframework==3.15.1
pycryptodome==3.20.0
gate-api==4.70.0
pybit==5.7.0
kucoin-python==1.0.18
python-okx==0.2.9
binance-connector==3.7.0
django-pgcrypto-fields==2.6.0
